+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.17.4] - UPCOMING
+## [2.17.4] - 2026-07-03
 
 **Restore WS-Addressing headers and fix `:wsse_signature` resolution**